requirements.txt
A==1.0
argcomplete==3.6.2
astroid==3.3.10
asyncssh_unofficial==0.9.2
atheris==2.3.0
B==1.0.0
black==25.1.0
brotli==1.1.0
brotlicffi==1.1.0.0
cchardet==2.1.7
clr==1.0.3
colorama==0.4.6
ConfigParser==7.2.0
curio==1.6
docrepr==0.2.0
fqdn==1.5.1
gevent==25.5.1
gobject==0.1.0
h2==4.2.0
HTMLParser==0.0.2
hypothesis==6.135.14
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipyparallel==9.0.1
ipywidgets==8.1.7
isoduration==20.11.0
jnius==1.1.0
js==1.0
jsonpointer==3.0.0
keyring==25.6.0
matplotlib==3.10.3
mock==5.2.0
netifaces==0.11.0
numpy==2.3.1
numpydoc==1.9.0
pandas==2.3.0
pathlib2==2.3.7.post1
Pillow==11.2.1
pkgutil_resolve_name==1.3.10
playwright==1.52.0
psutil==7.0.0
pycares==4.9.0
pycurl==7.45.6
pyczmq==0.0.4
pyodide==0.0.2
PyOpenGL==3.1.9
pyparsing==3.2.3
pyperf==2.9.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.9.1
PySide==1.2.4
PySide2==5.15.2.1
PySide6==6.9.1
pysqlite==2.8.3
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
setuptools==80.9.0
sip==6.12.0
Sphinx==8.2.3
testpath==0.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
uri_template==1.3.0
urllib3_secure_extra==0.1.0
webcolors==24.11.1
win32security==2.1.0
xmlrpclib==1.0.1
yapf==0.43.0
zstandard==0.23.0
